3-react-search-agent/pyproject.toml
[project]
name = "3-react-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
]
